releases/v0.1/requirements_ROMAIN.txt
astroid==2.0.1
future==0.16.0
huepy==0.9.8.1
iso8601==0.1.12
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pigments==1.6
prompt-toolkit==2.0.3
Pygments==2.2.0
pyserial==3.4
PyYAML==3.13
six==1.11.0
wcwidth==0.1.7
wrapt==1.10.11
